Move plone CSV data hook into its own file
- just a little layout improvement
